# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Pedidos",
        "description": "API para gestionar pedidos y ventas, incluyendo la creación de pedidos, consulta de productos y entrega.",
        "contact": {
            "email": "soporte@tusistema.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "api/administradores/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener todos los clientes",
                "operationId": "f3a3167835e99ea8cd9a8fd7cba39011",
                "responses": {
                    "200": {
                        "description": "Lista de clientes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Cliente"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los clientes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Crear un nuevo cliente",
                "operationId": "5b475c15ed0dabadb35f160579db180f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "cedula",
                                    "telefono",
                                    "password",
                                    "nickname"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del cliente",
                                        "type": "string"
                                    },
                                    "cedula": {
                                        "description": "Cédula del cliente",
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "description": "Teléfono del cliente",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del cliente",
                                        "type": "string"
                                    },
                                    "nickname": {
                                        "description": "Nombre de usuario del cliente",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cliente creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente creado exitosamente"
                                        },
                                        "cliente": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error al crear el cliente"
                    }
                }
            }
        },
        "api/administradores/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener un cliente por ID",
                "operationId": "dff3c48527cff24dbda26646a56560fe",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "500": {
                        "description": "Error al obtener el cliente"
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Actualizar los datos de un cliente",
                "operationId": "eb34569e49f9a8b81450acaa71f73790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "cedula",
                                    "telefono",
                                    "estado",
                                    "nickname"
                                ],
                                "properties": {
                                    "nombre": {
                                        "description": "Nombre del cliente",
                                        "type": "string"
                                    },
                                    "cedula": {
                                        "description": "Cédula del cliente",
                                        "type": "string"
                                    },
                                    "telefono": {
                                        "description": "Teléfono del cliente",
                                        "type": "string"
                                    },
                                    "estado": {
                                        "description": "Estado del cliente",
                                        "type": "string",
                                        "enum": [
                                            "ACTIVO",
                                            "INACTIVO"
                                        ]
                                    },
                                    "nickname": {
                                        "description": "Nombre de usuario del cliente",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente actualizado exitosamente"
                                        },
                                        "cliente": {
                                            "$ref": "#/components/schemas/Cliente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error al actualizar el cliente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Eliminar un cliente",
                "operationId": "a0240a61c0b63b24624fb75fb437587d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del cliente a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cliente eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Cliente no encontrado"
                    },
                    "500": {
                        "description": "Error al eliminar el cliente"
                    }
                }
            }
        },
        "api/administradores/clientes/login": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Iniciar sesión de un cliente",
                "operationId": "ac93b6a938133a1e294752a0f95dcae2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nickname",
                                    "password"
                                ],
                                "properties": {
                                    "nickname": {
                                        "description": "Nombre de usuario del cliente",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Contraseña del cliente",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Cliente autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cliente autenticado"
                                        },
                                        "token": {
                                            "description": "Token JWT generado",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error al autenticar al cliente"
                    }
                }
            }
        },
        "api/administradores/clientes/profile": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Obtener el perfil del cliente autenticado",
                "operationId": "465351e61f67c2fb2f0259b9e217c59c",
                "responses": {
                    "200": {
                        "description": "Datos del cliente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "api/administradores/clientes/logout": {
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Cerrar sesión del cliente",
                "operationId": "86281d33ae82b7e7b6acf1acd1aa7629",
                "responses": {
                    "200": {
                        "description": "Cliente desconectado"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Cliente": {
                "required": [
                    "id",
                    "nombre",
                    "cedula",
                    "telefono",
                    "nickname",
                    "estado",
                    "password",
                    "fecha_creacion"
                ],
                "properties": {
                    "id": {
                        "description": "ID del cliente",
                        "type": "integer"
                    },
                    "estado": {
                        "description": "Estado del cliente",
                        "type": "string",
                        "enum": [
                            "ACTIVO",
                            "INACTIVO"
                        ]
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creación del cliente",
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecha_actualizacion": {
                        "description": "Fecha de última actualización del cliente",
                        "type": "string",
                        "format": "date-time"
                    },
                    "nombre": {
                        "description": "Nombre del cliente",
                        "type": "string"
                    },
                    "cedula": {
                        "description": "Cédula del cliente",
                        "type": "string"
                    },
                    "telefono": {
                        "description": "Teléfono del cliente",
                        "type": "string"
                    },
                    "password": {
                        "description": "Contraseña del cliente (no se debe enviar en respuestas)",
                        "type": "string"
                    },
                    "nickname": {
                        "description": "Nombre de usuario del cliente",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Utiliza un token de autorización Bearer para autenticar las solicitudes.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}